LayoutTests/fast/dom/replace-children-mutation-events-expected.txt

This tests mutating DOM in mutation event listerns during replaceChilden.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parent.replaceChildren(span) threw exception NotFoundError: The object can not be found here..
PASS span.replaceChildren(oldParent, em, "text") did not throw exception.
PASS span.childNodes.length is 1
PASS span.childNodes[0] is oldParent
PASS span.textContent is ""
PASS successfullyParsed is true

TEST COMPLETE
some
